+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Devel
 
 * Bugfix: correctly import AnVIL groups when the service account is both a member and an admin of the group.
+* Bugfix: Fix failing ManagedGroup.anvil_audit_memership when the service account is not directly an admin of the group (but is via membership in another group).
 
 ## 0.16 (2023-06-01)